shopsmart-recsys/requirements.txt
fastapi
uvicorn[standard]
torch
pandas
numpy
faiss-cpu
pyyaml
pydantic